Progress
Can now load settings from the database and dynamically create a rough GUI for these settings, placing their JComponents in the correct tab according to its SettingCategory. For Settings of type Spinner, one can (and MUST) define which values are valid for this Setting in the database through a json array "validValues" in the data field; these potential values then become the model of the Spinner, with the Setting's value as the selected value.