...ends in failure
Failed attempt to make Settings be added to their SettingCategory as a JPanel with a JLabel for the Setting and the Setting itself. This was to ensure that the JLabel and the Setting would always be on the same line, and never be wrapped into separate lines - Config will instead be rewritten to allow this.